Revert "Update tests to use Doze DeviceConfig flags."
This change went into internal after R branched, and did not
go to AOSP. Revert it in mainline-prod in order to ensure that
CtsHostsideNetworkTests builds.

These tests are developed in internal, not AOSP, so do not need
to automerge to mainline-prod.

Bug: 167645754
Test: treehugger
Ignore-AOSP-First: needed to sync mainline-prod with AOSP
Merged-In: Ib0ac49609e444a53a6fee4575f5078e15f364eef
Change-Id: I2b7d9855b81be5ef968a3de72bf9f7e7225075e1