Use NET_CAPABILITY_NOT_VCN_MANAGED from ConstantsShim.
This is required to build in mainline-prod, which does not have
access to in-progress S APIs.

The Merged-In tag in this CL is from aosp/1604974, which is an
arbitrary CL that is already in sc-dev, with the purpose of
preventing this CL from reaching sc-dev.

Bug: 167645754
Ignore-AOSP-First: need to fix build
Test: m NetworkStack CtsNetTestCasesLatestSdk CtsHostsideNetworkTests CtsHostsideNetworkTestsApp
Change-Id: If40ebfadbf6d3ee86a58a55711fbc146a8599fdb
Merged-In: I5cd32efd0f87c823cd2d0a2fa3a95a83093fb6f9