Define PRIVATE_DNS_MODE_OPPORTUNISTIC locally
The type of ConnectivityManager#PRIVATE_DNS_MODE_OPPORTUNISTIC
has changed from String to int, but the String definition is still
needed to update to Settings.Global.PRIVATE_DNS_MODE, so the
simplest way is to define one locally.

Bug: 185311744
Test: atest CtsNetTestCases
Change-Id: Iafcd861714d8aca44cede658ed630f9d5afd5e59
Ignore-AOSP-First: In the same TOPIC, there is a merge conflict in
DevicePolicyManagerService.java in the internal branch, so it's needed
to merge the code into internal branch to fix the merge conflict.