Remove TEST_MAPPING duplicate imports
The NetworkStack import is already in the duplicate import
clause below it.

This was added by a wrong automerge during the last project
move, and was fixed in some of the downstream branches but
not all. This change should make imports identical in all branches.

Test: treehugger needs to test TEST_MAPPING
Merged-In: I7b5fd61cd551c7010f5b8ceabbfdd04f30f648dd
Merged-In: I71723d3459c14644140b81901dac972b2a746094
Change-Id: I6cbd31506329722d7fa5195a833d7acb4e4787a0