Restore private dns host name after testing
The helper method in CtsNetUtils will not restore the private
dns hostname back if the device private dns mode is not strict
mode. It does not cause function break since the host
configuration is useful only when the private dns mode is in
strict mode. But tests should restore the setting back to its
old state to prevent break other tests or erase setttings.

Test: atest android.net.cts.ConnectivityManagerTest and
      check if the private dns server configuration is
      changed or not after testing
Change-Id: I7c85ddac7306c7c3eeac84679d96c4cfb11bd875