Remove NetworkCapabilities#combine* tests
This is already in android12-tests-dev, but is necessary in this
branch as it is used for regression testing using internal test
configs which are not available on android12-tests-dev.
This change only includes the test changes, leaving out the
NetworkCapabilities changes which removed some unused code.

Original change: https://android-review.googlesource.com/c/platform/packages/modules/Connectivity/+/1919248

This removes methods in NetworkCapabilities that have been
removed from the mainline module in aosp/1919248, and as such
the code no longer has these private methods (including on S-V2
because it has the module) and we can't test for them.

For this reason these tests should be removed, as they won't
pass on a device with a recent module, and they have served
no purpose for a long time anyway.

Change-Id: I66a25bb9f165c29532c13ceed49000bc1e6a7587
Test: FrameworksNetTests
      m
Bug: 244271253
Bug: 218802331
Merged-In: Ia69c20afa16c2153839891f6e33331caa9da33ff
Merged-In: I58462f6d0303803ffd43333b67865c57edef39a3
Merged-In: Id5cca1972e115aeeb1061594dda07e39147630ac