Add android.net to proguard flags
Keep classes in android.net, as a followup change will have
some utility classes in that package in service-connectivity.

Ideally only "entry point" classes would be listed in the flags,
and the optimizer would figure out what is unreachable and
can be stripped out; but to be safe for now, list all classes.

Ignore-AOSP-First: Proguard rules not yet cherry-picked
Bug: 204830222
Change-Id: I45954f3bd23e84bacfd0f94875e10687e3e9cbc9
Test: m